# ENH: sparse: optimize CSR outer indexing