Takes advantage of a common case: "outer" fancy indexing.

Using the micro-benchmark from #5265 on my dev machine, before this PR:

    R[sel_rows][:, sel_cols]  =>  100 loops, best of 3: 3.65 ms per loop
    R[sel_rows[:,None], sel_cols]  =>  1 loops, best of 3: 1.05 s per loop

using this PR:

    R[sel_rows][:, sel_cols]  =>  100 loops, best of 3: 3.53 ms per loop
    R[sel_rows[:,None], sel_cols]  =>  100 loops, best of 3: 3.51 ms per loop

The `ismatrix` change fixes a bug in the logic. Previously, the `issequence(t) and issequence(t[0])` check would always fail, because `issequence` checks that its first element is a scalar.