@pv 2015-09-18
Maybe add test for the ismatrix logic issue?