@jakevdp 2015-09-18
Thanks for this – so the old version of ``ismatrix()`` would always return False? Any ideas where this may be causing problems in the current release?