@perimosocordiae 2015-09-18
I couldn't find any usages of `ismatrix()` in the master branch (using the Github repo search, at least). I haven't checked it's history, though.

I agree that a test is in order. I'll add one shortly.