@perimosocordiae 2015-09-18
Looks like `ismatrix` was added by 6814b7f4b298e80eaa3b0e09d8421dbfddebb32e, and it worked fine then because [`issequence` was less strict](https://github.com/scipy/scipy/blob/6814b7f4b298e80eaa3b0e09d8421dbfddebb32e/scipy/sparse/sputils.py#L126).

The commit that broke the logic is 0e9891e35ef17d578a9b781faf3f7daa9ebaba1b. All usages of `ismatrix` had been removed by that point, though, so the bug slipped by unnoticed.