@perimosocordiae 2015-09-18
Test added and Travis is green.