@jakevdp 2015-09-18
Thanks for looking into that.