@jakevdp 2015-09-18
Looked over this – looks good. In case anyone is curious, the main behavior affected here is already explicitly unit tested: https://github.com/scipy/scipy/blob/master/scipy/sparse/tests/test_base.py#L2338