@jakevdp 2015-09-18
So this is a weird corner case, but if you do
```
sel_rows = np.matrix([[0, 1],
                      [1, 0]])
sel_cols = np.matrix([[0], [1]])
```
then the ``isintlike(x[0][0])`` returns true, and the new logic will ignore all but the first column of ``rows``. I'm not sure at the moment the best way to fix that...