@perimosocordiae 2015-09-21
@jakevdp That case is handled, in a somewhat strange way, by the second half of the condition:

```python
In [27]: sel_rows
Out[27]:
matrix([[0, 1],
        [1, 0]])

In [28]: sel_rows[0]
Out[28]: matrix([[0, 1]])

In [29]: len(sel_rows[0]) == 1
Out[29]: True

In [30]: from scipy.sparse.sputils import isintlike

In [31]: isintlike(sel_rows[0][0])
Out[31]: False

In [32]: sel_rows[0][0]
Out[32]: matrix([[0, 1]])
```

I'm willing to add a test case, add a comment, or both. Thoughts?