@jakevdp 2015-09-21
Sorry – re-reading my comment it doesn't make sense.

I meant to say that ``isintlike(sel_rows[0][0])`` always returns False for a ``np.matrix``, regardless of the shape of the matrix, and thus the fast code will not be used.