@perimosocordiae 2015-09-21
I was surprised to find out that you were wrong about `np.matrix` not taking the fast path:

```python
In [36]: row = np.matrix([[0],[1]])

In [37]: row[0][0]
Out[37]: matrix([[0]])

In [38]: isintlike(row[0][0])
Out[38]: True
```