@jakevdp 2015-09-22
What?!?!? My intuition has misled me once again... :smile: