@jakevdp 2015-09-22
Now that you've disabused me of my confusion, I think this is good to go! Thanks for the contribution!